requirements.txt
slack-sdk>=3.19.0
python-dotenv>=1.0.0
streamlit>=1.28.0
